+++ src/games/snake/src/const/const.ts
@@ -1,0 +1,5 @@
+export let CONST = {
+  SCORE: 0,
+  HIGHSCORE: 0,
+  FIELD_SIZE: 8
+};