+++ src/games/snake/src/scenes/boot-scene.ts
@@ -1,0 +1,11 @@
+export class BootScene extends Phaser.Scene {
+  constructor() {
+    super({
+      key: 'BootScene'
+    });
+  }
+
+  update(): void {
+    this.scene.start('MainMenuScene');
+  }
+}