+++ CHANGELOG.md
@@ -31,6 +31,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Refactor `candy-crush`
 - Refactor `coin-runner`
 - Refactor `flappy-bird`
+- Refactor `snake`
 
 ---